Removed HTTPResources package -> Changed to resources Moved HTTPClient to WeatherRoute package as it's specific to that HTTP call
